requirements.txt
assemblyai>=0.34
python-dotenv>=1.0
fastapi>=0.120.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.20
tenacity>=9.0.0

yt-dlp>=2025.1.1
imageio-ffmpeg>=0.4.9
